+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# 2.30.0-0 / Unreleased
+# 2.30.1-0 / Unreleased
+  * Apply and enforce 'prettier' codestyle
+  * [Fixed] #894: Fancytree assertion failed: scrollParent should be a simple element or `window`, not document or body.
+  * [Fixed] #896 _requireExtension: order managment
+  * [Changed] #897 Set font for table extension
+
+# 2.30.0 / 2018-09-02
   * [Changed] ext-edit trigger 'clickActive' now only triggers if no modifier keys
     (shift, meta, control, ...) are pressed.<br>
     Trigger 'shift+click' now only triggers if no other modifier key (control, ...)
@@ -7,11 +13,13 @@
   * [Added] tree.activateKey(key, opts) now has an `opts` argument
   * [Added] `nodata` option (bool, string, or callback)
   * [Added] ext-table `mergeStatusColumns` option
+  * [Added] new method `tree.enable(flag)`
+  * [Added] new method `tree.expandAll(flag, opts)`
+  * [Added] new methods `tree.setOption(name, value)` and `tree.getOption(name)`
   * [Fixed] ES6 import dependency on jquery for jquery.fancytree.ui-deps.js
   * [Fixed] #863 setActive() sometimes does not scroll node into view
   * [Fixed] #877 postProcess may now also return the object form `{..., children: []}`
   * [Fixed] #884 ReferenceError: jQuery is not defined at _simpleDeepMerge
-    (before, only an array `[...]` was handled correctly).
   * [Fixed] autoScroll, node.scrollIntoView(), and .makeVisible() now work for tables as well.
   * [Changed] #879 Rename ext-debug to ext-logger
     (jquery.fancytree.debug.js => jquery.fancytree.logger.js)